Rearrange everything
