+++ MyWork/patch_config.py
@@ -1,0 +1,97 @@
+from torch import optim
+from patch_functions import *
+from loss_functions import *
+from dataset_functions import *
+
+
+class loss1(object):
+    """
+    Using the loss max_prob_class
+    """
+
+    def __init__(self):
+        """
+        Set the defaults.
+        """
+        self.patch_name = 'max_prob_class'
+
+        self.img_size = 640
+
+        # Dataloader informations
+        self.img_dir = '/home/andread98/yolov3/MyWork/dataset'    # train images location
+        self.lab_dir = '/home/andread98/yolov3/MyWork/dataset/yolo-labels'  # train labels location
+        self.batch_size = 6
+        self.n_epochs = 20
+        self.max_lab = 30
+
+        # Loss function 
+        self.loss_function = max_prob_class(0)
+
+        # Patch functions
+        # self.list_function_tile = [Tile_Creator_Circle, Tile_Creator_Ellipse, Tile_Creator_Square, Tile_Creator_Rectangle, Tile_Creator_Triangle, Tile_Creator_Trapezoid]
+        self.list_function_tile = [Tile_Creator_Double_Circle, Tile_Creator_Double_Ellipse, Tile_Creator_Double_Square, Tile_Creator_Double_Rectangle, Tile_Creator_Double_Triangle, Tile_Creator_Double_Trapezoid]
+        # self.list_function_params = [Params_Creator_Circle, Params_Creator_Ellipse, Params_Creator_Square, Params_Creator_Rectangle, Params_Creator_Triangle, Params_Creator_Trapezoid]
+        self.mask_function = Mask_Creator
+        self.dim_tile = 16
+        self.dim_patch = 640
+        self.mul_fact = 4 # multiplying factor, how big can the tile be wrt the dim_tile 
+        self.rotation_mode = 2
+        # 0: no rotation of the tile
+        # 1: rotation of a multiple of 90
+        # 2: random rotation. Attention, then the radius has to be at most sart(2)/2
+
+        # Optimizer informations
+        self.start_learning_rate = 0.03
+        self.scheduler_factory = lambda x: optim.lr_scheduler.ReduceLROnPlateau(x, 'min', patience=50)
+        # reduce learning rate when a metric has stopped learning (keras??)
+        # In min mode, lr will be reduced when the quantity monitored has stopped decreasing;
+        # in max mode it will be reduced when the quantity monitored has stopped increasing. Default: ‘min’.
+
+
+class loss2(loss1):
+    """
+    Using the loss max_prob_class2
+    """
+    def __init__(self):
+        """
+        Change stuff...
+        """
+        super().__init__()
+
+        self.patch_name = 'max_prob_class2'
+        self.loss_function = max_prob_class2(0)
+
+
+class loss3(loss1):
+    """
+    Using the loss max_prob_obj
+    """
+    def __init__(self):
+        """
+        Change stuff...
+        """
+        super().__init__()
+
+        self.patch_name = 'max_prob_obj'
+        self.loss_function = max_prob_obj()
+
+class loss4(loss1):
+    """
+    Using the loss new_loss_tprob
+    """
+    def __init__(self):
+        """
+        Change stuff...
+        """
+        super().__init__()
+
+        self.patch_name = 'new_loss_tprob'
+        self.loss_function = new_loss_tprob()
+
+
+patch_configs = {
+    "max_prob_class": loss1,
+    "max_prob_class2": loss2,
+    "max_prob_obj": loss3,
+    "new_loss_tprob": loss4
+}
